drm/msm/dsi-staging: Send notification when going from or to LP mode
When panel power mode is changed from or to low-power mode then no
notification is sent because notifications are sent only during
bridge enable or disable and this is not case for transitions
like ON -> LPx, LPx -> ON or LPx -> LPy. For such cases we need
to send a notification from dsi_display_set_power().

Change-Id: I3a10e097b4516579be0bf2e513262dc8705c12e4
Signed-off-by: Ivan Vecera <ivan@cera.cz>