drm/msm/dsi-staging: Add callback to provide alpha value for FOD dimming layer
Add callback to provide alpha value for FOD dimming layer that will be
used to mask flickering caused by HBM mode that is enabled temporarily
when user put finger down on FOD.

Luminance represented by backlight value is linear but alpha values used
for dimming layer not and they need to be gamma corrected. This requires
usage of pow() function and FPU. This is not possible in kernel so
device has to provide in DTS a look-up table that maps current backlight
level to value that eliminates a difference between current backlight
level and HBM backlight level.

Co-authored-by: Demon000 <demonsingur@gmail.com>
Signed-off-by: Ivan Vecera <ivan@cera.cz>
Change-Id: I63e44bbb5712996d9b69a01169424a8920839275